Moving tests into common directory to share between mysql and mysql2
